Move src/mitsubishi2mqtt.* -> src/main.*
